+++ docs/conf.py
@@ -23,7 +23,6 @@ extensions = [
     "sphinx_copybutton",
     "myst_parser",
     "nbsphinx",
-    "IPython.sphinxext.ipython_console_highlighting",
 ]
 
 intersphinx_mapping = {
@@ -63,7 +62,7 @@ html_sidebars = {
     ]
 }
 
-exclude_patterns = ["_build", "Thumbs.db", ".DS_Store", "**.ipynb_checkpoints"]
+exclude_patterns = ["_build", "Thumbs.db", ".DS_Store"]
 
 # -- Options for HTML output -------------------------------------------------